\name{as.integer.hashFactor}
\alias{as.integer.hashFactor}
\title{Convert a hashFactor to integer}
\usage{
  \method{as.integer}{hashFactor} (x, ...)
}
\arguments{
  \item{x}{hashFactor to be coerced to integer}

  \item{...}{ignored}
}
\description{
  Convert a hashFactor to integer
}
\author{
  Jeffrey D. Allen \email{Jeffrey.Allen@UTSouthwestern.edu}
}

