\name{as.integer.hashFactors}
\alias{as.integer.hashFactors}
\title{Convert a hashFactor to integer}
\usage{
  as.integer.hashFactors(x, ...)
}
\arguments{
  \item{x}{hashFactor to be coerced to integer}
}
\description{
  Convert a hashFactor to integer
}
\author{
  Jeffrey D. Allen \email{Jeffrey.Allen@UTSouthwestern.edu}
}

