\name{cast_type}
\alias{cast_type}
\title{Convert data type names}
\usage{
  cast_type(type)
}
\arguments{
  \item{type}{The JSON Table Schema type name.}
}
\value{
  The R type name.
}
\description{
  Convert the type names used in JSON Table Schema to their
  R equivalents.
}
\author{
  Jeffrey D. Allen \email{Jeffrey.Allen@UTSouthwestern.edu}
}

