\name{print.dataPackage}
\alias{print.dataPackage}
\title{Print a dataPackage}
\usage{
  \method{print}{dataPackage} (x, ...)
}
\arguments{
  \item{x}{The data package as read from
  \link{read_data_package}.}

  \item{...}{Ignored. Provided to be consistent with
  signature of global print function.}
}
\description{
  Print a list of class \code{dataPackage} in a clean way.
}
\author{
  Jeffrey D. Allen \email{Jeffrey.Allen@UTSouthwestern.edu}
}